apirestio/requirements.txt
beautifulsoup4==4.8.0
Django==2.2.5
django-admin-generator==2.2.0
django-admin-interface==0.11.2
django-filter==2.2.0
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphene-django-extras==0.4.6
Pillow==6.2.0
requests==2.22.0
django-silk
